+++ src/main/java/enterprises/orbital/evexmlapi/eve/IConquerableStation.java
@@ -2,9 +2,10 @@ package enterprises.orbital.evexmlapi.eve;
 
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 
-@JsonSerialize(as = IConquerableStation.class)
+@JsonSerialize(
+    as = IConquerableStation.class)
 public interface IConquerableStation {
-  public int getCorporationID();
+  public long getCorporationID();
 
   public String getCorporationName();
