+++ src/main/java/enterprises/orbital/impl/evexmlapi/eve/FactionStat.java
@@ -1,22 +1,24 @@
 package enterprises.orbital.impl.evexmlapi.eve;
 
-public abstract class FactionStat implements FacWarStat{
-	private int factionID;
-	private String factionName;
+public abstract class FactionStat implements FacWarStat {
+  private long   factionID;
+  private String factionName;
 
-	public int getFactionID() {
-		return factionID;
-	}
+  public long getFactionID() {
+    return factionID;
+  }
 
-	public void setFactionID(int factionID) {
-		this.factionID = factionID;
-	}
+  public void setFactionID(
+                           long factionID) {
+    this.factionID = factionID;
+  }
 
-	public String getFactionName() {
-		return factionName;
-	}
+  public String getFactionName() {
+    return factionName;
+  }
 
-	public void setFactionName(String factionName) {
-		this.factionName = factionName;
-	}
+  public void setFactionName(
+                             String factionName) {
+    this.factionName = factionName;
+  }
 }
